backend/pyproject.toml
[tool.poetry]
authors = ["cristian-rincon <cristian.o.rincon.b@gmail.com>"]
description = ""
license = "MIT"
name = "vote-app"
version = "0.1.0"

[tool.poetry.dependencies]
Django = "^4.0.2"
Markdown = "^3.3.6"
PyYAML = "^6.0"
dj-database-url = "^0.5.0"
django-extensions = "^3.1.5"
django-filter = "^21.1"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
python = "^3.9"
uritemplate = "^4.1.1"
whitenoise = "^6.0.0"

[tool.poetry.dev-dependencies]
pytest-cov = "^3.0.0"
pytest-django = "^4.5.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "voteapp.settings"
addopts = "-p no:warnings --strict-markers --no-migrations --reuse-db"
norecursedirs = ".venv old_tests"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[tool.coverage.run]
branch = true
omit = ["*/tests/*", "*/migrations/*", "*/urls.py", "*/settings/*", "*/wsgi.py", "manage.py", "fabfile.py", "*/settings.py", "*/asgi.py", "*__init__*"]
source = ["."]
[tool.coverage.report]
show_missing = true
